// src/components/Auth/libs/enums.ts
export enum Lables {
    FirstName = 'Имя',
    LastName = 'Фамилия',
    Email = 'Email',
    Company = 'Название компании',
    Password = 'Введите пароль',
    SubmitPassword = 'Подтвердите пароль',
    OldPassword = 'Старый пароль'
}

export enum Placeholders {
    FirstName = 'Введите ваше имя',
    LastName = 'Введите вашу фамилию',
    Email = 'Введите ваш email',
    Company = 'Введите название компании',
    Password = 'Введите ваш пароль',
    SubmitPassword = 'Подтвердите ваш пароль',
    OldPassword = 'Введите текущий пароль'
}

export enum InputNames {
    FirstName = 'first_name',
    LastName = 'last_name',
    Email = 'email',
    Company = 'company',
    Password = 'password',
    SubmitPassword = 're_password',
    OldPassword = 'old_password'
}

export enum Autocompletes {
    FirstName = 'given-name',
    LastName = 'family-name',
    Email = 'email',
    Company = 'organization',
    NewPassword = 'new-password',
    CurrentPassword = 'current-password'
}

export enum ErrorMessages {
    INCORRECT_DATA = 'Введены некорректные данные',
    REQUIRED_FILED = 'Поле обязательно для заполнения',
    NOT_MATCH_PASSWORD = 'Пароль не совпадает'
}